refactor: abstract Hit, Running, Finished, AfterInitDraw
