+++ src/main/java/blackjack/domain/participant/state/Finished.java
@@ -1,15 +1,12 @@
 package blackjack.domain.participant.state;
 
 import blackjack.domain.carddeck.Card;
-import java.util.ArrayList;
 import java.util.List;
 
-public abstract class Finished implements State {
-
-    private final List<Card> cards;
+public abstract class Finished extends AfterInitDraw {
 
     public Finished(final List<Card> cards) {
-        this.cards = cards;
+        super(cards);
     }
 
     @Override
@@ -20,11 +17,6 @@ public abstract class Finished implements State {
     @Override
     public State stay() {
         throw new IllegalStateException();
-    }
-
-    @Override
-    public List<Card> cards() {
-        return new ArrayList<>(this.cards);
     }
 
     @Override
