+++ src/main/java/blackjack/domain/participant/state/InitDraw.java
@@ -1,6 +1,7 @@
 package blackjack.domain.participant.state;
 
 import blackjack.domain.carddeck.Card;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -9,7 +10,7 @@ public class InitDraw {
     public static final int BLACKJACK = 21;
 
     public static State draw(final Card firstCard, final Card secondCard) {
-        List<Card> cards = Arrays.asList(firstCard, secondCard);
+        List<Card> cards = new ArrayList<>(Arrays.asList(firstCard, secondCard));
         if (isBlackjack(cards)) {
             return new Blackjack(cards);
         }
