+++ src/main/java/blackjack/domain/participant/state/Running.java
@@ -1,0 +1,16 @@
+package blackjack.domain.participant.state;
+
+import blackjack.domain.carddeck.Card;
+import java.util.List;
+
+public abstract class Running extends AfterInitDraw {
+
+    public Running(final List<Card> cards) {
+        super(cards);
+    }
+
+    @Override
+    public boolean isFinished() {
+        return false;
+    }
+}
